requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pyperclip>=1.8.2
python-dotenv
selenium
webdriver_manager
Pillow>=10.0.0
imagehash
packaging>=23.0
